Refactor request to trade configuration for token
Instead of using an `APNSConfiguration` object, now it takes a token and the configuration is applied by the client.

This makes more sense because the `APNSRequest` always needs to have a token – the token is an important part of the request; whereas the configuration is not – it dictates where to deliver the request but nothing about it.