Load dotenv in index.php
I'm not sure whether this is required or there's something missing in my
setup.

Even after creating `.env` locally, `php index.php` failed saying that
the init parameters for `APNSConfiguration` were booleans instead of
strings, that is, the environment variables couldn't be found.

Explicitly loading dotenv as per the documentation solved the problem
for me.

See
https://github.com/vlucas/phpdotenv/tree/c2da5440047bbaf8ed8dc29a0df300e88b5cb50b#usage