Add examples folder to list of folders to lint and fix issues
Most of the changes were made by `phpcbf`.

I'm not sure about the change from `date` to `gmdate`, the rationale
provided by PHPCS was:

> date() is affected by runtime timezone changes which can cause
> date/time to be incorrectly displayed. Use gmdate() instead.

I suppose it's a safe change to make, given this is only an example file
anyways.

I decided to disable the warnings on global functions and function name
not using snake casing for the same reason as above, this is an example
file.