Fix a flaky APNSSoundTest
If the sound level is `1.0`, we don’t need to serialize an APNSSound to an object – we can just serialize it to a string because `1.0` is the default. This caused a broken test, because it would serialize correctly 10% of the time in a way the test didn’t account for.